Adjust size of popup window to reach its target dimensions
Store the adjustments for the next time it's opened.
Select the first item in the popup window by default, so you can press the shortcut once and release the modifier to toggle between tabs.
Hide the window on blur when not in dev mode.
Set activeTabs to [] when the popup closes, so that if we refresh with dev tools open it won't error.
Get rid of isNaN in init-tabs.js.