Use previous/next tab commands to move the selection in the popup
Handle modifySelected message from the background to move the selection up or down.
Get the current settings in showWindow() so that the list gets rendered using the latest.
Add an onMessage listener in the popup regardless if it's rendered in a popup window.
Keep track of the popup and menu ports separately, so opening the menu doesn't cause the window to get closed and reopened.
Refactor tab navigation into navigateTabs(), and get rid of tabChangedFromCommand global.
Open the popup port with a name based on the window mode.
Close any open popup windows when the extension restarts.
Filter the popup from closed tabs.
Tweak line wrapping in chrome-store-description.txt.