Make onMessage listener not async
Making the handler async caused bugs when the popup and menu were open at the same time.  Not clear why it worked with only one open, but with both, the async responses were never sent.
Always send a response from onMessage to the app, to avoid "The message port closed before a response was received." errors.  sendMessage() in the app wrapped all calls in a promise, so it always needs a response.